Split the sufficiency judge into a package
`sufficiency_judge.py` was 545 lines with two of its three stages written, and
stage C plus verdict derivation and the panel runner would have roughly doubled
it. Splitting before stage C means stage C gets written into its own file
rather than moved out of a larger one later.

    src/eval/sufficiency/
        __init__.py   the criterion, the protocol, the layout. Exports nothing.
        models.py     dataclasses and literals — no LLM import
        llm.py        build_judge_llm — the only ai_common touchpoint
        stage_a.py    decompose
        stage_b.py    answer blind, span_is_verbatim
        judge.py      the driver; stage C and the panel go here

A pure move. Both prompts are byte-identical across it by AST comparison
(STAGE_A_INSTRUCTIONS sha 2b152da2, STAGE_B_INSTRUCTIONS sha f27fb440), the
pydantic schemas are unchanged, and the set of top-level definitions is the
same before and after — nothing lost, nothing added. Two dead imports (`re`,
`sys`) did not survive, having been unused.

Isolating `ai_common` in `llm.py` is the point of the shape, not tidiness: it
pulls langchain → transformers → torch, the same 8.34s that split `llm_config`
out of `config`. Keeping `models` clear of it means the dataclasses, and the
verdict derivation still to be written over them, import for nothing.

That property was claimed and then measured, and the measurement failed: a
convenience re-export of the stages in `__init__` made `import
src.eval.sufficiency.models` cost **7.68s**, because Python runs a package's
`__init__` before any submodule of it. The `__init__` now exports nothing at
all — there is no caller outside the package yet, so any public surface today
would be a guess — and the same import costs **0.07s** and does not reach
torch. The file says so, with the number, since the next convenience import
would silently undo it.

One consequence: the probe harness is now `python -m src.eval.sufficiency.judge`.

Also repointed to the new modules: the design document's code references, which
now name `module.py:symbol` rather than line ranges, per the convention in
docs/design/README.md; and two entries in todo.md.
